Moved redux directory to sit in src directory.  Made the redux store available to the content
