+++ src/content/components/DomView.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { Routes, Route } from "react-router-dom";
 import { Provider } from "react-redux";
-import { store } from "../../options/redux/store";
+import { store } from "../../redux/store";
 import { HashRouter as Router } from "react-router-dom";
 // import "../assets/tailwind.css";
 import DataInterface from "../../options/components/DataInterface";
