+++ src/content/index.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
+import { Provider } from "react-redux";
+import { store } from "../redux/store";
 // import "../assets/tailwind.css";
 import Content from "./content";
 
@@ -9,10 +11,14 @@ const init = () => {
   if (!appContainer) {
     throw new Error("can't find appContainer");
   }
-  
+
   document.body.appendChild(appContainer);
   const root = createRoot(appContainer);
   console.log(appContainer);
-  root.render(<Content />);
+  root.render(
+    <Provider store={store}>
+      <Content />
+    </Provider>
+  );
 };
 init();
